backend/requirements.txt
fastapi
uvicorn
psycopg2
sqlalchemy
alembic
pydantic
faker
python-dotenv